Remove constructor from Location
* src/main/java/is/hi/hbv501g1/hidden_pearls/location/Location.java:
  Remove the constructor.

Spring seems to require a default/noargs constructor to exists, and it
isn't any more hassle to create a default constructed Location and use
the setters instead of passing arguments to a constructor.  Especially
since the only time we should be creating Location objects instead of
retrieving them from Spring should be the creation of new Locations.